Adding support for AFHDS 3 and SAMD21 architecture
The channel parsing routine is updated to enable the ADHFS 3 protocol
supporting up to 18 channels.  This is done by interleaving channels
15-18 in the unused bits between channels 1-12.

In AFHDS 2A the extra bytes are always 0 so it is safe to replace the
old parser with this new parser.

The SAMD21 ISR timer is added to enable support for SAMD21 architectures
such as the MKR Wifi 1010 or the Arduino Zero.